pyproject.toml
[project]
name = "vors-ting"
version = "0.1.0"
description = "Multi-Agent Critical Feedback Scaffolding"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "numpy>=1.24",
    "scikit-learn>=1.2",
    "sentence-transformers>=2.2",
    "bandit>=1.9.3",
    "safety>=3.7.0",
    "mccabe>=0.7.0",
    "pyright>=1.1.408",
    "mkdocs>=1.6.1",
    "mkdocstrings>=1.0.3",
    "semgrep>=1.151.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "pytest-asyncio>=0.21",
    "pytest-xdist>=3.0",
    "pytest-benchmark>=4.0",
    "pytest-timeout>=2.0",
    "pytest-randomly>=4.0",
    "pytest-sugar>=1.0",
    "pytest-html>=4.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocstrings[python]>=0.25",
]

[tool.uv]
managed = true
preview = true

[project.scripts]
vors = "vors_ting.cli:app"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP", "YTT", "S", "B", "A", "BLE", "C4", "C90", "DTZ", "FA", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "SLOT", "T20", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["D203", "D213"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "SLF001", "PLR2004", "PLC0415"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=vors_ting --cov-report=term-missing"
markers = [
    "slow: marks tests as slow (deselect with '-m not slow')",
    "integration: marks integration tests (deselect with '-m not integration')",
]

[tool.coverage.run]
source = ["vors_ting"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.mkdocs]
theme = "readthedocs"
plugins = ["mkdocstrings"]

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__", "**/.pytest_cache", "**/.mypy_cache"]
venvPath = "."
venv = ".venv"
reportMissingImports = "warning"
reportMissingModuleSource = "warning"
pythonVersion = "3.13"
